+++ lib/fqn/fqn.go
@@ -1,4 +1,4 @@
-package util
+package fqn
 
 import (
 	"errors"
