+++ lib/fqn/fqn_test.go
@@ -1,4 +1,4 @@
-package util
+package fqn
 
 import (
 	"strings"
